use quantized model

+++ model.js
@@ -1,27 +1,41 @@
 import wavEncoder from 'wav-encoder';
 import ort from 'onnxruntime-node';
 import fs from 'node:fs/promises';
+import fsSync from 'node:fs';
 import phonetisaurus from 'phonetisaurus';
 
 function intersperse(array, something) {
   if (array.length < 2) { return array }
-  var result = [], i = 0, l = array.length
+
+  const result = []
+  let i = 0;
+  let l = array.length;
+
   if (typeof something == 'function') {
     for (; i < l; i++) {
       if (i !== 0) { result.push(something()) }
       result.push(array[i])
     }
-  }
-  else {
+  } else {
     for (; i < l; i++) {
       if (i !== 0) { result.push(something) }
       result.push(array[i])
     }
   }
 
-  result.unshift(something)
-  result.push(something)
+  result.unshift(something);
+  result.push(something);
   return result
+}
+
+if (!fsSync.existsSync("g2p.fst")) {
+  console.error("'g2p.fst' file not found")
+  process.exit(1);
+}
+
+if (!fsSync.existsSync("G_60000.int8.onnx")) {
+  console.error("'G_60000.int8.onnx' file not found")
+  process.exit(1);
 }
 
 phonetisaurus.FS.writeFile("/model.fst", await fs.readFile("g2p.fst"));
@@ -32,7 +46,7 @@ export async function tts(text) {
   const phonemes = phonemizer.phoneticize(text, 1, 500, 10, false, false, 0.0)[0].join(" ");
   const ids = (phonemes + ".").split('').map(t => tokenizer[t] || tokenizer[' ']);
   const inputs = [intersperse(ids, 0)];
-  const session = await ort.InferenceSession.create('G_60000.onnx');
+  const session = await ort.InferenceSession.create('G_60000.int8.onnx');
   const wrap = v => new ort.Tensor('int64', BigInt64Array.from(v.flat().map(x => BigInt(x))), [v.length, v[0].length]);
   const result = await session.run({
     input: wrap(inputs),
@@ -47,5 +61,6 @@ export async function tts(text) {
     sampleRate,
     channelData,
   });
+
   return Buffer.from(audioBuffer)
 }
